Moved the test folder and added some more documentation
