+++ src/test/java/xyz/thepathfinder/android/PathfinderTest.java
@@ -1,4 +1,4 @@
-package xyz.thepathfinder.android.test;
+package xyz.thepathfinder.android;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
@@ -6,8 +6,6 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import xyz.thepathfinder.android.Cluster;
-import xyz.thepathfinder.android.Pathfinder;
 
 import java.io.IOException;
 import java.net.URI;
