+++ src/test/java/xyz/thepathfinder/android/TestMessager.java
@@ -1,4 +1,4 @@
-package xyz.thepathfinder.android.test;
+package xyz.thepathfinder.android;
 
 import java.util.logging.Logger;
